+++ app/(tabs)/presyo.tsx
@@ -232,7 +232,7 @@ export default function PresyoScreen() {
         contentContainerStyle={{ paddingTop: headerHeight ?? 0 }}
         keyboardShouldPersistTaps="handled"
         refreshControl={
-          <RefreshControl refreshing={refreshing} onRefresh={onRefresh} tintColor="#386641" />
+          <RefreshControl refreshing={refreshing} onRefresh={onRefresh} tintColor="#386641" colors={['#386641']} />
         }
         onScroll={Animated.event([{ nativeEvent: { contentOffset: { y: scrollY } } }], { useNativeDriver: false })}
         scrollEventThrottle={16}