+++ app/help.tsx
@@ -161,7 +161,7 @@ export default function HelpScreen() {
 
       <ScrollView
         contentContainerStyle={{ padding: 16, paddingBottom: insets.bottom + 32 }}
-        refreshControl={<RefreshControl refreshing={refreshing} onRefresh={onRefresh} tintColor="#386641" />}
+        refreshControl={<RefreshControl refreshing={refreshing} onRefresh={onRefresh} tintColor="#386641" colors={['#386641']} />}
       >
         {tab === 'faq' ? (
           faqsLoading ? (