+++ app/my-reports.tsx
@@ -102,7 +102,7 @@ export default function MyReportsScreen() {
 
       <ScrollView
         contentContainerStyle={{ padding: 16, paddingBottom: insets.bottom + 32 }}
-        refreshControl={<RefreshControl refreshing={refreshing} onRefresh={onRefresh} tintColor="#386641" />}
+        refreshControl={<RefreshControl refreshing={refreshing} onRefresh={onRefresh} tintColor="#386641" colors={['#386641']} />}
       >
         {isLoading ? (
           <>